+++ sdks/typescript/src/v1/client/features/crons.ts
@@ -52,42 +52,6 @@ export class CronClient {
     return str;
   }
 
-  private async getWorkflowId(
-    workflow: string | Workflow | BaseWorkflowDeclaration<any, any>
-  ): Promise<string> {
-    const str = (() => {
-      if (typeof workflow === 'string') {
-        return workflow;
-      }
-
-      if (typeof workflow === 'object' && 'name' in workflow) {
-        return workflow.name;
-      }
-
-      if (typeof workflow === 'object' && 'id' in workflow) {
-        if (!workflow.id) {
-          throw new Error('Workflow ID is required');
-        }
-        return workflow.id;
-      }
-
-      throw new Error(
-        'Invalid workflow: must be a string, Workflow object, or WorkflowDefinition object'
-      );
-    })();
-
-    if (!isValidUUID(str)) {
-      console.log('getting workflow id', str);
-      const wf = await this.workflows.get(str);
-      if (!wf) {
-        throw new Error('Invalid workflow ID: must be a valid UUID');
-      }
-      return wf.metadata.id;
-    }
-
-    return str;
-  }
-
   /**
    * Creates a new Cron workflow.
    * @param workflow - The workflow identifier or Workflow object.
@@ -148,8 +112,7 @@ export class CronClient {
     const { workflow, ...rest } = query;
 
     if (workflow) {
-      const workflowId = await this.getWorkflowId(workflow);
-      console.log('workflowId', workflowId);
+      const workflowId = await this.workflows.getWorkflowIdFromName(workflow);
       rest.workflowId = workflowId;
     }
